Moved state checking and transition checking into new StateMonitor class
